rust backend: try to run 'rustfmt --edition 2018' on generated code
The main benefit of this is that it yields better error messages for
users of the generated code if rustc refers to a span within it, e.g.

```
cargo run
   Compiling humblegen v0.3.0 (/home/cs/49nord/humblegen/humblegen-rs/humblegen)
   Compiling humblegen-example-project v0.1.0 (/home/cs/49nord/humblegen/humblegen-rs/example)
error[E0046]: not all trait items implemented, missing: `get_monster_id`
   --> src/api_impl.rs:14:1
    |
14  | impl protocol::Godzilla for MonsterApiImpl {
    | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ missing `get_monster_id` in implementation
    |
   ::: /home/cs/49nord/humblegen/humblegen-rs/example/target/debug/build/humblegen-example-project-3e5b23a77b58eb99/out/protocol.rs:175:5
    |
175 |     async fn get_monster_id(&self, id: String) -> Response<Result<(), MonsterError>>;
    |     --------------------------------------------------------------------------------- `get_monster_id` from trait
```

where line 175 can be displayed as a single line as opposed to one
gigantic line that contains the entire generated code and thus cannot be
displayed properly.